Refactor graphql api
