+++ spec/graphql_queries/fetch_pr_request.json.erb
@@ -1,1 +1,1 @@
-{"query":"query RedmineGithub__GithubAPI__Repository<%= variables[:repository_id] %>Client__GET_PR_QUERY($number: Int!, $repo_owner: String!, $repo_name: String!) {\n  repository(owner: $repo_owner, name: $repo_name) {\n    pullRequest(number: $number) {\n      number\n      title\n      state\n      mergeable\n      mergedAt\n      mergeStateStatus\n    }\n  }\n}","variables":{"number":<%= variables[:pull_request_number] %>,"repo_owner":"<%= variables[:repo_owner] %>","repo_name":"<%= variables[:repo_name] %>"},"operationName":"RedmineGithub__GithubAPI__Repository<%= variables[:repository_id] %>Client__GET_PR_QUERY"}
+{"query":"query RedmineGithub__GithubAPI__Graphql__Repository<%= variables[:repository_id] %>Client__FETCH_PR_QUERY($number: Int!, $repo_owner: String!, $repo_name: String!) {\n  repository(owner: $repo_owner, name: $repo_name) {\n    pullRequest(number: $number) {\n      number\n      title\n      state\n      mergeable\n      mergedAt\n      mergeStateStatus\n    }\n  }\n}","variables":{"number":<%= variables[:pull_request_number] %>,"repo_owner":"<%= variables[:repo_owner] %>","repo_name":"<%= variables[:repo_name] %>"},"operationName":"RedmineGithub__GithubAPI__Graphql__Repository<%= variables[:repository_id] %>Client__FETCH_PR_QUERY"}
