+++ spec/lib/github_api/graphql_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require File.expand_path('../rails_helper', __dir__)
+require File.expand_path('../../rails_helper', __dir__)
 
-RSpec.describe RedmineGithub::GithubAPI do
+RSpec.describe RedmineGithub::GithubAPI::Graphql do
   before do
     graphql_mock(
       request: graphpl_json_for(:load_schema_request),
